Extract feature-family plugin builders
Add AbstractFeaturePluginBuilder and three concrete subclasses covering
the project-feature plugin shapes:

- FeaturePluginBuilder: single binding, real build model
- NoBuildModelFeaturePluginBuilder: single binding with BuildModel.None
  (Kotlin rendering uses the reified-style binding)
- MultiTargetFeaturePluginBuilder: multiple binding targets sharing a
  generic BaseApplyAction (Java only — Kotlin rendering not currently
  exercised by any fixture)

Move the shared helpers maybeDeclareBindingModifiers and
generateCustomServices to AbstractPluginBuilder. Move the feature-only
helpers (parent type resolution, feature service injection, mapping/
display delegators) to AbstractFeaturePluginBuilder.
