+++ platforms/core-configuration/declarative-dsl-provider/src/testFixtures/groovy/org/gradle/features/internal/builders/FeaturePluginBuilder.groovy
@@ -1,0 +1,165 @@
+/*
+ * Copyright 2026 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.features.internal.builders
+
+import org.gradle.api.provider.ProviderFactory
+import org.gradle.features.annotations.BindsProjectFeature
+import org.gradle.features.binding.BuildModel
+import org.gradle.features.binding.ProjectFeatureApplicationContext
+import org.gradle.features.binding.ProjectFeatureApplyAction
+import org.gradle.features.binding.ProjectFeatureBinding
+import org.gradle.features.binding.ProjectFeatureBindingBuilder
+import org.gradle.features.file.ProjectFeatureLayout
+import org.gradle.features.registration.TaskRegistrar
+
+/**
+ * Generates a {@code @BindsProjectFeature}-annotated plugin with a single binding to a definition
+ * (or definition build model) and a real (non-{@code BuildModel.None}) build model.
+ */
+class FeaturePluginBuilder extends AbstractFeaturePluginBuilder {
+
+    @Override
+    protected String renderJava() {
+        def modifiers = maybeDeclareBindingModifiers()
+        def implType = primaryDefinition.implementationClassName
+            ? ".withUnsafeDefinitionImplementationType(${primaryDefinition.implementationClassName}.class)"
+            : ""
+        def bmImplType = primaryDefinition.buildModelFullImplementationClassName
+            ? ".withBuildModelImplementationType(${primaryDefinition.buildModelFullImplementationClassName}.class)"
+            : ""
+
+        def parentType = getParentTypeForApplyAction()
+        def buildModelType = primaryDefinition.buildModelFullPublicClassName
+
+        return """
+            package ${packageName};
+
+            import org.gradle.api.Plugin;
+            import org.gradle.api.Project;
+            import ${BindsProjectFeature.class.name};
+            import ${ProjectFeatureBindingBuilder.class.name};
+            import static ${ProjectFeatureBindingBuilder.class.name}.bindingToTargetDefinition;
+            import ${ProjectFeatureBinding.class.name};
+            import ${ProjectFeatureApplyAction.class.name};
+            import ${ProjectFeatureApplicationContext.class.name};
+            import ${BuildModel.class.name};
+
+            @${BindsProjectFeature.class.simpleName}(${pluginClassName}.Binding.class)
+            public class ${pluginClassName} implements Plugin<Project> {
+
+                static class Binding implements ${ProjectFeatureBinding.class.simpleName} {
+                    @Override public void bind(${ProjectFeatureBindingBuilder.class.simpleName} builder) {
+                        builder.${bindingMethodName}(
+                            "${name}",
+                            ${primaryDefinition.className}.class,
+                            ${bindingTypeClassName}.class,
+                            ${pluginClassName}.ApplyAction.class
+                        )
+                        ${implType}${bmImplType}${modifiers};
+                    }
+                }
+
+                static abstract class ApplyAction implements ${ProjectFeatureApplyAction.class.simpleName}<${primaryDefinition.className}, ${buildModelType}, ${parentType}> {
+                    @javax.inject.Inject public ApplyAction() { }
+
+                    ${generateJavaFeatureApplyActionServices()}
+
+                    @Override
+                    public void apply(${ProjectFeatureApplicationContext.class.simpleName} context, ${primaryDefinition.className} definition, ${buildModelType} model, ${parentType} parent) {
+                        System.out.println("Binding ${primaryDefinition.className}");
+                        System.out.println("${name} model class: " + model.getClass().getSimpleName());
+                        System.out.println("${name} parent model class: " + context.getBuildModel(parent).getClass().getSimpleName());
+
+                        ${buildModelMappingForLanguage()}
+
+                        getTaskRegistrar().register("print${primaryDefinition.className}Configuration", task -> {
+                            task.doLast(t -> {
+                                ${displayDefinitionValuesForLanguage()}
+                                ${displayModelValuesForLanguage()}
+                            });
+                        });
+                    }
+                }
+
+                @Override
+                public void apply(Project project) {
+
+                }
+            }
+        """
+    }
+
+    @Override
+    protected String renderKotlin() {
+        def modifiers = maybeDeclareBindingModifiers()
+        def parentType = getParentTypeForApplyAction()
+        def buildModelType = primaryDefinition.buildModelFullPublicClassName
+
+        return """
+            package ${packageName}
+
+            import org.gradle.api.Plugin
+            import org.gradle.api.Project
+            import ${BindsProjectFeature.class.name}
+            import ${ProjectFeatureBindingBuilder.class.name}
+            import ${ProjectFeatureBinding.class.name}
+            import ${ProjectFeatureApplyAction.class.name}
+            import ${ProjectFeatureApplicationContext.class.name}
+            import javax.inject.Inject
+
+            @${BindsProjectFeature.class.simpleName}(${pluginClassName}.Binding::class)
+            class ${pluginClassName} : Plugin<Project> {
+
+                class Binding : ${ProjectFeatureBinding.class.simpleName} {
+                    override fun bind(builder: ${ProjectFeatureBindingBuilder.class.simpleName}) {
+                        builder.${bindingMethodName}("${name}", ${primaryDefinition.className}::class.java, ${bindingTypeClassName}::class.java, ${pluginClassName}.ApplyAction::class.java)${modifiers}
+                    }
+                }
+
+                abstract class ApplyAction @Inject constructor() : ${ProjectFeatureApplyAction.class.simpleName}<${primaryDefinition.className}, ${buildModelType}, ${parentType}> {
+
+                    @get:Inject
+                    abstract val taskRegistrar: ${TaskRegistrar.class.name}
+
+                    @get:Inject
+                    abstract val projectFeatureLayout: ${ProjectFeatureLayout.class.name}
+
+                    @get:Inject
+                    abstract val providerFactory: ${ProviderFactory.class.name}
+
+                    override fun apply(context: ${ProjectFeatureApplicationContext.class.name}, definition: ${primaryDefinition.className}, model: ${buildModelType}, parent: ${parentType}) {
+                        println("Binding ${primaryDefinition.className}")
+                        println("${name} model class: " + model::class.simpleName)
+                        println("${name} parent model class: " + context.getBuildModel(parent)::class.simpleName)
+
+                        ${buildModelMappingForLanguage()}
+
+                        taskRegistrar.register("print${primaryDefinition.className}Configuration") { task ->
+                            task.doLast { _ ->
+                                ${displayDefinitionValuesForLanguage()}
+                                ${displayModelValuesForLanguage()}
+                            }
+                        }
+                    }
+                }
+
+                override fun apply(project: Project) {
+                }
+            }
+        """
+    }
+}
